-will avoid deleting log file when classify called with --overwrite -log file set to always debug level output -dumped args file and configs to debug log

+++ voila/classify.py
@@ -170,8 +170,10 @@ def run_classifier():
 
     if os.listdir(config.directory):
         if config.overwrite:
+            log_file_relpath = config.logger.split('/')[-1] if config.logger else None
             for _file in os.listdir(config.directory):
-                os.remove(os.path.join(config.directory, _file))
+                if _file != log_file_relpath:
+                    os.remove(os.path.join(config.directory, _file))
         else:
             voila_log().critical(
                 "Files already present in %s; not running classifier "
